Moved Random class to own file, changed MathHelper to be partial
